backend/requirements.txt
# Web Framework & Server
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.10.0
python-multipart>=0.0.12

# AI & RAG Orchestration
langchain>=1.2.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-pinecone>=0.2.0

# Database & Vectors
pinecone-client>=5.0.0

# Document Processing
pypdf>=5.0.0

# Utilities
python-dotenv>=1.0.1
tiktoken>=0.8.0